SudokuGenerator_5.1
Completed (?) file to generate Sudoku puzzles in 2 different ways (a simple pattern, and a backtracking algorithm).  Also includes a number of helper functions for a UI to use.
Main change from version _4 is the addition of the SudokuBoard object which stores the data of a Sudoku board in an easily accessed way.